--- a/types/overflow.go
+++ b/types/overflow.go
@@ -49,6 +49,16 @@ func AddDuration(a time.Duration, b time.Duration) (time.Duration, error) {
 	return a + b, nil
 }
 
+// SubDuration subtracts time.Duration a with b and returns time.Duration if no overflow error.
+func SubDuration(a time.Duration, b time.Duration) (time.Duration, error) {
+	if (a > 0 && b < 0 && math.MaxInt64-a < -b) ||
+		(a < 0 && b > 0 && math.MinInt64-a > -b) ||
+		(a == 0 && b == math.MinInt64) {
+		return 0, ErrOverflow.GenWithStackByArgs("BIGINT", fmt.Sprintf("(%d, %d)", a, b))
+	}
+	return a - b, nil
+}
+
 // AddInteger adds uint64 a and int64 b and returns uint64 if no overflow error.
 func AddInteger(a uint64, b int64) (uint64, error) {
 	if b >= 0 {