<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Use go template to generate vectorized  11 SUBTIME functions from #12176 

- builtinSubDatetimeAndDurationSig

- builtinSubDatetimeAndStringSig

- builtinSubDurationAndDurationSig

- builtinSubDurationAndStringSig

- builtinSubStringAndDurationSig

- builtinSubStringAndStringSig

- builtinSubDateAndDurationSig

- builtinSubDateAndStringSig

- builtinSubTimeDateTimeNullSig

- builtinSubTimeStringNullSig

- builtinSubTimeDurationNullSig

### What is changed and how it works?



    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDatetimeAndDurationSig-VecBuiltinFunc-4               3922            302150 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDatetimeAndDurationSig-NonVecBuiltinFunc-4            3393            361119 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDatetimeAndStringSig-VecBuiltinFunc-4                  624           1899306 ns/op           94497 B/op       5814 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDatetimeAndStringSig-NonVecBuiltinFunc-4               624           2202165 ns/op           94320 B/op       5814 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDurationAndDurationSig-VecBuiltinFunc-4             108002             10695 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDurationAndDurationSig-NonVecBuiltinFunc-4           30878             38418 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDurationAndStringSig-VecBuiltinFunc-4                  768           1525616 ns/op           91830 B/op       5634 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDurationAndStringSig-NonVecBuiltinFunc-4               765           1602095 ns/op           91497 B/op       5634 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubStringAndDurationSig-VecBuiltinFunc-4                  510           2658512 ns/op          185895 B/op       9004 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubStringAndDurationSig-NonVecBuiltinFunc-4               465           2423924 ns/op          185699 B/op       9004 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubStringAndStringSig-VecBuiltinFunc-4                    382           3301695 ns/op          286293 B/op      14953 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubStringAndStringSig-NonVecBuiltinFunc-4                 367           3186760 ns/op          285774 B/op      14953 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDateAndDurationSig-VecBuiltinFunc-4                   2694            411262 ns/op           96050 B/op       3939 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDateAndDurationSig-NonVecBuiltinFunc-4                2307            477730 ns/op           96051 B/op       3939 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDateAndStringSig-VecBuiltinFunc-4                      544           2288155 ns/op          202358 B/op      10537 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDateAndStringSig-NonVecBuiltinFunc-4                   484           2272191 ns/op          201554 B/op      10537 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeDateTimeNullSig-VecBuiltinFunc-4              13504646                83.4 ns/op             0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeDateTimeNullSig-NonVecBuiltinFunc-4              76623             15796 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeStringNullSig-VecBuiltinFunc-4                  368790              3064 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeStringNullSig-NonVecBuiltinFunc-4               147936              7918 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeDurationNullSig-VecBuiltinFunc-4              14170660                90.3 ns/op             0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeDurationNullSig-NonVecBuiltinFunc-4             111211              9843 ns/op               0 B/op          0 allocs/op




### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test