@SunRunAway 2019-11-25
@SilvaXiang builtinSubXXXAndYYYSig may be resolved by go generate, could you take a look at https://docs.google.com/document/d/17XdYo_EwYaWANqbjS3i3OVSz3oLlovU2fIo1bRx4YqA/edit#