@SilvaXiang 2019-11-27
> @SilvaXiang builtinSubXXXAndYYYSig may be resolved by go generate, could you take a look at https://docs.google.com/document/d/17XdYo_EwYaWANqbjS3i3OVSz3oLlovU2fIo1bRx4YqA/edit#

请问一定只能通过Go generate方式改吗？能不能一个一个往builtin_time_generated_vec.go里面加需要修改的函数？