@SilvaXiang 2019-11-27
> > > @SilvaXiang builtinSubXXXAndYYYSig may be resolved by go generate, could you take a look at https://docs.google.com/document/d/17XdYo_EwYaWANqbjS3i3OVSz3oLlovU2fIo1bRx4YqA/edit#
> > 
> > 
> > 请问一定只能通过Go generate方式改吗？能不能一个一个往builtin_time_generated_vec.go里面加需要修改的函数？
> 
> Hi, since builtinAddXXXAndYYYSig is already templated, we'd like to template builtinSubXXXAndYYYSig too, so that it's better for us to maintain in the future.
> 
> builtin_time_generated_vec.go is auto-generated by expression/generator/time_vec.go, each time we run `go generate`, the file will be regenerated, so it should not be modified by hands.
> 
> You may want to take a look at expression/generator/time_vec.go, try to template `sub` logic with `add`.

Hi，I have finished 9 functions by `go generate`. Waiting for your review. Thanks.