@SilvaXiang 2019-11-28
@SunRunAway Hey,I have finished 9 functions by `go generate`. Waiting for your review. Thanks.