@SilvaXiang 2019-11-29
@qw4990 Hey,I have finished 9 functions by go generate. Waiting for your review. Thanks.