@SunRunAway 2019-12-04
It seems that you've miss two functions (builtinSubStringAndDurationSig and builtinSubStringAndStringSig).