@SunRunAway 2019-12-05
Please consider this comment https://github.com/pingcap/tidb/pull/13691#discussion_r354163318